Error displayed while fetching rollout revisions details (#10132)
